+++ module/apmgin/middleware.go
@@ -94,6 +94,7 @@ func (m *middleware) handle(c *gin.Context) {
 			setContext(&e.Context, c, body)
 			e.Send()
 		}
+		c.Writer.WriteHeaderNow()
 		tx.Result = apmhttp.StatusCodeResult(c.Writer.Status())
 
 		if tx.Sampled() {
@@ -115,10 +116,8 @@ func setContext(ctx *apm.Context, c *gin.Context, body *apm.BodyCapturer) {
 	ctx.SetFramework("gin", gin.Version)
 	ctx.SetHTTPRequest(c.Request)
 	ctx.SetHTTPRequestBody(body)
-	if c.Writer.Written() {
-		ctx.SetHTTPStatusCode(c.Writer.Status())
-		ctx.SetHTTPResponseHeaders(c.Writer.Header())
-	}
+	ctx.SetHTTPStatusCode(c.Writer.Status())
+	ctx.SetHTTPResponseHeaders(c.Writer.Header())
 }
 
 // Option sets options for tracing.