+++ module/apmhttp/recovery.go
@@ -21,6 +21,8 @@ type RecoveryFunc func(
 // The returned RecoveryFunc will report recovered error to Elastic APM
 // using the given Tracer, or apm.DefaultTracer if t is nil. The
 // error will be linked to the given transaction.
+//
+// If headers have not already been written, a 500 response will be sent.
 func NewTraceRecovery(t *apm.Tracer) RecoveryFunc {
 	if t == nil {
 		t = apm.DefaultTracer
@@ -35,7 +37,7 @@ func NewTraceRecovery(t *apm.Tracer) RecoveryFunc {
 	) {
 		e := t.Recovered(recovered)
 		e.SetTransaction(tx)
-		setContext(&e.Context, req, resp, body)
+		SetContext(&e.Context, req, resp, body)
 		e.Send()
 	}
 }