+++ CHANGELOG.md
@@ -22,6 +22,7 @@
  - `ELASTIC_APM_SANITIZE_FIELD_NAMES` and `ELASTIC_APM_IGNORE_URLS` now use wildcard matching (#260)
  - Changed top-level package name to "apm", and canonical import path to "go.elastic.co/apm" (#202)
  - module/apmrestful: introduce emicklei/go-restful instrumentation (#270)
+ - Fix panic handling in web instrumentations (#273)
 
 ## [v0.5.2](https://github.com/elastic/apm-agent-go/releases/tag/v0.5.2)